requirements.txt
# Django
django
djangorestframework

# Web Scraping
linkedin_scraper
selenium
